Different formatting for Letter entries Fixes #4
