+++ app/components/letter-entry.module.css
@@ -1,0 +1,9 @@
+.letter {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+}
+
+.letterHeading {
+    align-self: center;
+}
