+++ app/components/letter-heading.module.css
@@ -1,0 +1,3 @@
+.headings {
+    align-self: flex-end;
+}
