+++ app/page.js
@@ -1,6 +1,6 @@
 import styles from './page.module.css'
 import React from 'react';
-import JournalEntry from './components/journal-entry';
+import LatestJournalEntry from './components/latest-journal-entry';
 
 
 export default function Home() {
@@ -14,7 +14,7 @@ export default function Home() {
       <p>Bram Stoker</p>
       <p>How these papers have been placed in sequence will be made manifest in the reading of them. All needless matters have been eliminated, so that a history almost at variance with the possibilities of later-day belief may stand forth as simple fact. There is throughout no statement of past things wherein memory may err, for all the records chosen are exactly contemporary, given from the standpoints and within the range of knowledge of those who made them.</p>
 
-      <JournalEntry api={api}></JournalEntry>     
+      <LatestJournalEntry api={api}></LatestJournalEntry>     
     </main>
   );
